@@ -1,53 +1,37 @@
-const path = require('path');
-var webpack = require('webpack');
-var CompressionPlugin = require('compression-webpack-plugin');
+const HtmlWebPackPlugin = require("html-webpack-plugin");
 
 module.exports = {
-    entry: {
-        app: [__dirname + '/src'],
-        "browse-drop-standalone": [__dirname + '/src/components/wrapper/standalone.js']
-    },
-    output: {
-        path: path.resolve(__dirname, "dist"),
-        filename: '[name].js',
-        publicPath: '/dist'
-    },
-    //devtool: "cheap-module-source-map",
-    devtool: "source-map",
-    //devtool: "cheap-module-eval-source-map",
     module: {
-        rules: [{
-            test: /\.css$/,
-            use: ['style-loader', 'css-loader']
-        }, {
-            test: /\.styl$/,
-            use: ['style-loader', 'css-loader', 'stylus-loader'],
-        }, {
-            test: /\.(js|jsx)$/,
-            exclude: /(node_modules|bower_components)/,
-            use: {
-                loader: 'babel-loader'
-            }
-        }, {
-            test: /\.styl$/,
-            use: [{
-                loader: "style-loader" // creates style nodes from JS strings
-            }, {
-                loader: "css-loader" // translates CSS into CommonJS
-            }, {
-                loader: "stylus-loader" // compiles Stylus to CSS
-            }]
-        }]
-    },
-    plugins: [],
-    resolve: {
-        alias: {
-            'react': path.join(__dirname, 'node_modules', 'react')
+      rules: [
+        {
+          test: /\.(js|jsx)$/,
+          exclude: /node_modules/,
+          use: {
+            loader: "babel-loader"
+          }
         },
-        extensions: ['.js', '.jsx', '.styl']
+        {
+          test: /\.html$/,
+          use: [
+            {
+              loader: "html-loader"
+            }
+          ]
+        },
+        {
+          test: /\.styl$/i,
+          use: [
+              'style-loader',
+              'css-loader',
+              'stylus-loader'
+          ]
+        }
+      ]
     },
-    externals: {
-        'react': 'React',
-        'react-dom': 'ReactDOM'
-    }
-};
+    plugins: [
+      new HtmlWebPackPlugin({
+        template: "./src/js/components/index.html",
+        filename: "./index.html"
+      })
+    ]
+  };
